+++ packages/koenig-lexical/src/nodes/CodeBlockNode.jsx
@@ -1,91 +1,15 @@
-import * as React from 'react';
-import CardContext from '../context/CardContext';
-import KoenigComposerContext from '../context/KoenigComposerContext.jsx';
+import React from 'react';
 import cleanBasicHtml from '@tryghost/kg-clean-basic-html';
 import generateEditorState from '../utils/generateEditorState';
 import {$generateHtmlFromNodes} from '@lexical/html';
-import {$getNodeByKey} from 'lexical';
-import {ActionToolbar} from '../components/ui/ActionToolbar.jsx';
 import {CodeBlockNode as BaseCodeBlockNode} from '@tryghost/kg-default-nodes';
-import {CodeBlockCard} from '../components/ui/cards/CodeBlockCard';
 import {ReactComponent as CodeBlockIcon} from '../assets/icons/kg-card-type-gen-embed.svg';
+import {CodeBlockNodeComponent} from './CodeBlockNodeComponent';
 import {KoenigCardWrapper, MINIMAL_NODES} from '../index.js';
-import {SnippetActionToolbar} from '../components/ui/SnippetActionToolbar.jsx';
-import {ToolbarMenu, ToolbarMenuItem, ToolbarMenuSeparator} from '../components/ui/ToolbarMenu.jsx';
 import {createEditor} from 'lexical';
-import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 
 // re-export here so we don't need to import from multiple places throughout the app
 export {INSERT_CODE_BLOCK_COMMAND} from '@tryghost/kg-default-nodes';
-
-function CodeBlockNodeComponent({nodeKey, captionEditor, captionEditorInitialState, code, language}) {
-    const [editor] = useLexicalComposerContext();
-    const {isEditing, setEditing, isSelected} = React.useContext(CardContext);
-    const {cardConfig} = React.useContext(KoenigComposerContext);
-    const [showSnippetToolbar, setShowSnippetToolbar] = React.useState(false);
-
-    const updateCode = (value) => {
-        editor.update(() => {
-            const node = $getNodeByKey(nodeKey);
-            node.setCode(value);
-        });
-    };
-
-    const updateLanguage = (value) => {
-        editor.update(() => {
-            const node = $getNodeByKey(nodeKey);
-            node.setLanguage(value);
-        });
-    };
-
-    const handleToolbarEdit = (event) => {
-        event.preventDefault();
-        event.stopPropagation();
-        setEditing(true);
-    };
-
-    return (
-        <>
-            <CodeBlockCard
-                captionEditor={captionEditor}
-                captionEditorInitialState={captionEditorInitialState}
-                code={code}
-                handleToolbarEdit={handleToolbarEdit}
-                isEditing={isEditing}
-                isSelected={isSelected}
-                language={language}
-                nodeKey={nodeKey}
-                updateCode={updateCode}
-                updateLanguage={updateLanguage}
-            />
-            <ActionToolbar
-                data-kg-card-toolbar="button"
-                isVisible={showSnippetToolbar}
-            >
-                <SnippetActionToolbar onClose={() => setShowSnippetToolbar(false)} />
-            </ActionToolbar>
-
-            <ActionToolbar
-                data-kg-card-toolbar="button"
-                isVisible={isSelected && !isEditing}
-            >
-                <ToolbarMenu>
-                    <ToolbarMenuItem dataTestId="edit-code-card" icon="edit" isActive={false} label="Edit" onClick={handleToolbarEdit} />
-                    <ToolbarMenuSeparator hide={!cardConfig.createSnippet} />
-                    <ToolbarMenuItem
-                        dataTestId="create-snippet"
-                        hide={!cardConfig.createSnippet}
-                        icon="snippet"
-                        isActive={false}
-                        label="Snippet"
-                        onClick={() => setShowSnippetToolbar(true)}
-                    />
-                </ToolbarMenu>
-            </ActionToolbar>
-        </>
-    );
-}
-
 export class CodeBlockNode extends BaseCodeBlockNode {
     // transient properties used to control node behaviour
     __openInEditMode = false;
@@ -97,7 +21,7 @@ export class CodeBlockNode extends BaseCodeBlockNode {
 
         const {_openInEditMode} = dataset;
         this.__openInEditMode = _openInEditMode || false;
-        
+
         // set up and populate nested editors from the serialized HTML
         this.__captionEditor = dataset.captionEditor || createEditor({nodes: MINIMAL_NODES});
         this.__captionEditorInitialState = dataset.captionEditorInitialState;
